Refractored, fixed a message_header import by moving the content of ddex.ddex into ddex.__init__ Allowed for python 27 and python 3+ versions of tkinter to be imported Added an alternative method for mkdir -p 56 / 82 unit tests pass
