+++ ddex/tests/test_file_parser.py
@@ -1,10 +1,14 @@
+import os
 import unittest
+
 from nose.tools import *
-from DDEXUI.file_parser import FileParser
+
+from ddexui.file_parser import FileParser
+
 
 def test_generator():
-    cases = ([ ("ddex/tests/resources/test.mp3", "dff9465befeb68d97cd6fd103547c464", "test.mp3", "MP3"), 
-                ("ddex/tests/resources/test.jpg", "55e031153f2c0d8c63e6bf7c9baa58ba", "test.jpg", "JPG")])
+    cases = ([ (os.path.join(os.path.dirname(__file__), "resources", "test.mp3"), "dff9465befeb68d97cd6fd103547c464", "test.mp3", "MP3"), 
+                (os.path.join(os.path.dirname(__file__), "resources", "test.jpg"), "55e031153f2c0d8c63e6bf7c9baa58ba", "test.jpg", "JPG")])
     for path, hash, name, extension in cases:
         yield check_file, path, hash, name, extension
 
@@ -18,7 +22,7 @@ class FileParserTests(unittest.TestCase):
 
     def setUp(self):
         self.subject = FileParser()
-        self.file_metadata = self.subject.parse("ddex/tests/resources/test.mp3")
+        self.file_metadata = self.subject.parse(os.path.join(os.path.dirname(__file__), "resources", "test.mp3"))
 
     def test_should_have_duration(self):
         self.assertEqual(self.file_metadata.duration, "PT0M4.000S")
@@ -32,7 +36,7 @@ class FileParserTests(unittest.TestCase):
 class ImageFileParserTests(unittest.TestCase):
     def setUp(self):
         self.subject = FileParser()
-        self.file_metadata = self.subject.parse("ddex/tests/resources/test.jpg")
+        self.file_metadata = self.subject.parse(os.path.join(os.path.dirname(__file__), "resources", "test.jpg"))
 
     def test_should_have_height(self):
         self.assertEqual(self.file_metadata.height, 500)
