+++ ddex/tests/test_party_repository.py
@@ -1,7 +1,6 @@
 import unittest
-import configparser
-from DDEXUI.ddex.party import *
-from DDEXUI.party_repository import *
+from ddex.party import *
+from ddexui.party_repository import *
 import sqlite3
