+++ ddex/tests/test_release_id.py
@@ -1,4 +1,4 @@
-from DDEXUI.ddex.release import ReleaseId
+from ddex.release import ReleaseId
 import unittest
 
 class ReleaseIdTests(unittest.TestCase):
