+++ ddex/tests/test_validates_against_ddex_schema.py
@@ -1,13 +1,15 @@
-import lxml.etree as ET
-from DDEXUI.ddex.ddex import DDEX
-from DDEXUI.ddex.release import Release, ReleaseId
-from DDEXUI.ddex.party import Party, PartyType
-from DDEXUI.ddex.deal import Deal
-from DDEXUI.ddex.resource import SoundRecording, Image
-from DDEXUI.ddex.message_header import MessageHeader
-from DDEXUI.file_parser import FileParser
 from datetime import date
+import os
 import unittest
+
+from ddex import DDEX
+from ddex.deal import Deal
+from ddex.party import Party, PartyType
+from ddex.release import Release, ReleaseId
+from ddex.resource import SoundRecording, Image
+from ddexui.file_parser import FileParser
+import lxml.etree as ET
+
 
 class DDEXSchemaValidation(unittest.TestCase):
     def test_created_ddex_files_validate_against_ddex_xsd(self):
@@ -51,10 +53,10 @@ class DDEXSchemaValidation(unittest.TestCase):
 
     def create_sound_recording(self):
         resource_reference = "A1"
-        resource = SoundRecording(resource_reference, "abc", "Bad", FileParser().parse("ddex/tests/resources/test.mp3"),"T1")
+        resource = SoundRecording(resource_reference, "abc", "Bad", FileParser().parse(os.path.join(os.path.dirname(__file__), "resources", "test.mp3")),"T1")
         return resource
     
     def create_image(self):
         image_resource_reference = "A2"
-        image_resource = Image(image_resource_reference, "abc", FileParser().parse("ddex/tests/resources/test.jpg"),"T2")
+        image_resource = Image(image_resource_reference, "abc", FileParser().parse(os.path.join(os.path.dirname(__file__), "resources", "test.jpg")),"T2")
         return image_resource
