+++ ddexui/batch_generator.py
@@ -1,6 +1,6 @@
-from os import makedirs
+import os
 from os import path
-from DDEXUI.ddex.ddex import generate_batch_id
+from ddex import generate_batch_id
 
 class BatchGenerator:
     def __init__(self, root_folder, batch_id):
@@ -12,5 +12,10 @@ class BatchGenerator:
         for builder in builders:
             ddex = builder.build()
             product_path = path.join(batch_path, builder.get_upc())
-            makedirs(product_path, exist_ok=True)
+            # Could use exist_ok in python 3+ but not available in 27
+            try: 
+                os.makedirs(product_path)
+            except OSError:
+                if not os.path.isdir(path):
+                    raise
             ddex.write(path.join(product_path, builder.get_upc() + ".xml"))
