+++ ddexui/file_parser.py
@@ -1,5 +1,5 @@
-from DDEXUI.ddex.file_metadata import *
-from mutagenx.mp3 import MP3
+from ddex.file_metadata import *
+from mutagen.mp3 import MP3
 from PIL import Image
 import hashlib
 import os
