+++ ddexui/release_window.py
@@ -1,17 +1,24 @@
-import tkinter.ttk as tk
+try:
+    import ttk as tk
+    from Tkinter import Toplevel
+    from tkFileDialog import askopenfilename
+except ImportError:
+    import tkinter.ttk as tk
+    from tk.tkinter import Toplevel
+    from tkinter.filedialog import askopenfilename
 from PIL import Image, ImageTk
-from tkinter.filedialog import askopenfilename
-from DDEXUI.ddex.release_builder import ReleaseBuilder
-from DDEXUI.ddex.validate import Validate
-from DDEXUI.inputs import *
-from DDEXUI.ddex.release import *
-from DDEXUI.deal_window import DealWindow
-from DDEXUI.file_parser import FileParser
-from DDEXUI.tkinterutil import showerrorbox
-from DDEXUI.resource_manager import ResourceManager
-from DDEXUI.product_service import ProductService
+from ddex.release_builder import ReleaseBuilder
+from ddex.validate import Validate
+from inputs import *
+from ddex.release import *
+from deal_window import DealWindow
+from file_parser import FileParser
+from tkinterutil import showerrorbox
+from resource_manager import ResourceManager
+from product_service import ProductService
 
-class ReleaseWindow(tk.tkinter.Toplevel):
+
+class ReleaseWindow(Toplevel):
     def __init__(self, frame):
         tk.tkinter.Toplevel.__init__(self, frame)
         self._release_builder = ReleaseBuilder()
