+++ requirements.txt
@@ -1,4 +1,6 @@
-mutagenx>=1.21
+mutagen>=1.21
 nose==1.3.0
 pillow==2.2.1
-lxml==3.2.0
+lxml==3.4.1
+enum34==1.0.4
+cx_Freeze==4.3.4
